src/python/requirements.txt
tensorflow
tensorflow-datasets
zensols.dbpg~=1.1.0
zensols.util~=1.13.0
